Basic new file structure  for function distribution
